+++ src/knowledge/sources/index.ts
@@ -1,7 +1,0 @@
-export * from "./builtin-product.js";
-export * from "./merge.js";
-export * from "./registry.js";
-export * from "./routing.js";
-export * from "./search.js";
-export * from "./types.js";
-export * from "./workspace.js";